requirements.txt
certifi==2017.4.17
chardet==3.0.4
gevent==1.2.2
greenlet==0.4.12
idna==2.5
pymongo==3.4.0
PyYAML==3.12
requests==2.18.1
urllib3==1.21.1
